WE'RE OUT OF ALPHA BABY
Finally, after weeks, and months of work, I can officially say, WE'RE RELEASED BABY!!! I'm so happy to finally say this, you don't know how happy I am really.

12:50: YOOO THIS WORKS!

+++ src/pages/engines/edit/Intake.jsx
@@ -149,12 +149,12 @@ const Intake = ({database}) => {
 										id,
 										values: {
 											...intake,
-											plenumVolume: parseInt(e.target.value),
+											plenumVolume: parseFloat(e.target.value),
 										},
 									});
 									setIntake({
 										...intake,
-										plenumVolume: parseInt(e.target.value),
+										plenumVolume: parseFloat(e.target.value),
 									});
 								}}
 							/>
@@ -176,12 +176,12 @@ const Intake = ({database}) => {
 										id,
 										values: {
 											...intake,
-											plenumCrossSectionArea: parseInt(e.target.value),
+											plenumCrossSectionArea: parseFloat(e.target.value),
 										},
 									});
 									setIntake({
 										...intake,
-										plenumCrossSectionArea: parseInt(e.target.value),
+										plenumCrossSectionArea: parseFloat(e.target.value),
 									});
 								}}
 							/>
@@ -203,12 +203,12 @@ const Intake = ({database}) => {
 										id,
 										values: {
 											...intake,
-											flowRate: parseInt(e.target.value),
+											flowRate: parseFloat(e.target.value),
 										},
 									});
 									setIntake({
 										...intake,
-										flowRate: parseInt(e.target.value),
+										flowRate: parseFloat(e.target.value),
 									});
 								}}
 							/>
@@ -230,12 +230,12 @@ const Intake = ({database}) => {
 										id,
 										values: {
 											...intake,
-											idleFlowRate: parseInt(e.target.value),
+											idleFlowRate: parseFloat(e.target.value),
 										},
 									});
 									setIntake({
 										...intake,
-										idleFlowRate: parseInt(e.target.value),
+										idleFlowRate: parseFloat(e.target.value),
 									});
 								}}
 							/>
@@ -285,12 +285,12 @@ const Intake = ({database}) => {
 										id,
 										values: {
 											...intake,
-											throttleGamma: parseInt(e.target.value),
+											throttleGamma: parseFloat(e.target.value),
 										},
 									});
 									setIntake({
 										...intake,
-										throttleGamma: parseInt(e.target.value),
+										throttleGamma: parseFloat(e.target.value),
 									});
 								}}
 							/>
